refac(bison): separated actions by folder and file
