+++ src/backend/support/logger.h
@@ -4,7 +4,7 @@
 #include <stdio.h>
 
 // Descomentar para loguear en modo DEBUG (con más detalle).
-#define DEBUG true
+//#define DEBUG true
 
 void Log(FILE * const stream, const char * prefix, const char * const format, const char * suffix, va_list arguments);
