+++ src/frontend/syntactic-analysis/actions/constant/constant.h
@@ -1,0 +1,7 @@
+#ifndef ACTION_CONSTANT_H
+#define ACTION_CONSTANT_H
+#include "../program.h"
+
+Constant* IntegerConstantGrammarAction(int value);
+
+#endif
