+++ src/frontend/syntactic-analysis/actions/declaration/declaration.h
@@ -1,0 +1,11 @@
+#ifndef ACTION_DECLARATION_H
+#define ACTION_DECLARATION_H
+
+#include "../program.h"
+
+
+// Declaration
+Declaration* DeclarationGrammarAction(int type, char* id);
+Declaration* VectorDeclarationGrammarAction(int type, char* id);
+
+#endif
