+++ src/frontend/syntactic-analysis/actions/property/property.h
@@ -1,0 +1,8 @@
+#ifndef ACTION_PROPERTY_H
+#define ACTION_PROPERTY_H
+
+#include "../program.h"
+
+Property* PropertyGrammarAction(PropertyType type, Type dataType);
+
+#endif
