+++ src/frontend/syntactic-analysis/actions/vector/vector.h
@@ -1,0 +1,8 @@
+#ifndef ACTION_VECTOR_H
+#define ACTION_VECTOR_H
+
+#include "../program.h"
+
+Vector* VectorGrammarAction(char* id, Factor* factor);
+
+#endif
