+++ src/frontend/syntactic-analysis/actions/while/while.h
@@ -1,0 +1,8 @@
+#ifndef ACTION_WHILE_H
+#define ACTION_WHILE_H
+
+#include "../program.h"
+
+While* WhileGrammarAction(Expression* expression, Block* block);
+
+#endif
